split headers (big step)
