+++ src/config.h
@@ -1,0 +1,20 @@
+/* Global configure
+ * Author: Xianjie Huang <hxj5@hku.hk>
+ */
+#ifndef CSP_CONFIG_H
+#define CSP_CONFIG_H
+
+#define DEBUG 0
+#define VERBOSE 1
+/* DEVELOP defined to 1 means some codes for future version of cellsnp will be included. */
+#define DEVELOP 0
+
+#define CSP_NAME "cellsnp-lite"
+#define CSP_VERSION "1.0.1"
+#define CSP_AUTHOR "hxj5<hxj5@hku.hk>"
+
+#define JF_ZIP_TYPE 2     // use bgzip as zip method for JFile
+
+typedef struct _gll_settings global_settings;
+
+#endif
