+++ src/kvec.h
