+++ src/mplp.h
@@ -1,564 +1,21 @@
-#ifndef CELL_SNP_UTIL_H
-#define CELL_SNP_UTIL_H
-
-#define DEBUG 0
-#define VERBOSE 1
-/* DEVELOP defined to 1 means some codes for future version of cellsnp will be included. */
-#define DEVELOP 0
-
-#define CSP_NAME "cellsnp-lite"
-#define CSP_VERSION "1.0.1"
-#define CSP_AUTHOR "hxj5<hxj5@hku.hk>"
-
-// zip method for csp_fs_t structure.
-#define CSP_FS_GZIP 1
-#define CSP_FS_BGZIP 2
-#define CSP_FS_ZIP_METHOD CSP_FS_BGZIP
+/* Pileup and MPileup operatoins API/routine
+ * Author: Xianjie Huang <hxj5@hku.hk>
+ */
+#ifndef CSP_MPLP_H
+#define CSP_MPLP_H
 
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include <math.h>
-#include <stdarg.h>
-#include "htslib/vcf.h"
 #include "htslib/sam.h"
 #include "htslib/kstring.h"
-#include "htslib/hts.h"
 #include "htslib/khash.h"
-#include "htslib/bgzf.h"
 #include "kvec.h"
-#include "general_util.h"
-
-typedef struct _gll_settings global_settings;
-
-/* 
-* BAM/SAM/CRAM File API 
-*/
-
-/*@abstract  Translate chr name to tid of bam/sam/cram.
-@param hdr   Pointer of sam_hdr_t structure.
-@param name  Chr name.
-@param s     Pointer of kstring_t.
-@return      Non-negative number if success, -1 for unrecognized reference seq name, -2 for unparsed hdr.
-
-@note        If the translation failed, this function will try "<name> - chr" if name starts with "chr", try
-             "chr + <name>" otherwise. 
- */
-static inline int csp_sam_hdr_name2id(sam_hdr_t *hdr, const char *name, kstring_t *s) {
-    int tid;
-    if ((tid = sam_hdr_name2tid(hdr, name)) < 0) {
-        if (tid < -1) { return tid; }
-        if (0 == strncmp(name, "chr", 3)) { return sam_hdr_name2tid(hdr, name + 3); }
-        else {
-            kputs("chr", s); kputs(name, s);
-            return sam_hdr_name2tid(hdr, ks_str(s));
-        }
-    } else { return tid; }
-}
-
-/*@abstract  Convert chrom name to be the same with the one in sam header. 
-@return      Pointer to ref/chrom name if success, NULL otherwise.
-@note        No need to free the returned char* pointer when success.
-*/
-static inline const char* csp_fmt_chr_name(const char *name, sam_hdr_t *hdr, kstring_t *s) {
-    int tid;
-    if ((tid = csp_sam_hdr_name2id(hdr, name, s)) < 0) { return NULL; }
-    else { return sam_hdr_tid2name(hdr, tid); }
-}
-
-/*@abstract   The two functions below convert raw cigar op/len value to real value.
-@param c      Raw cigar op/len value stored in bam1_t, can be an element of cigar array obtained by bam_get_cigar(b) [uint32_t].
-@return       An integer [int].
-*/
-#define get_cigar_op(c) ((c) & BAM_CIGAR_MASK)
-#define get_cigar_len(c) ((c) >> BAM_CIGAR_SHIFT)
-
-/*@abstract  Convert index of seq_nt16_str to the index (0-4) of A/C/G/T/G.
-@param i     Index in the seq_nt16_str for the base [int8_t].
-@return      Index in the 'ACGTN' for the base [int8_t].
- */
-#define seq_nt16_idx2int(i) (seq_nt16_int[i])
-
-/*@abstract  Convert a char (A/C/G/T/N) to index of seq_nt16_str/seq_nt16_int.
-@param c     The base char [int8_t].
-@return      Index in the seq_nt16_str for the base [int8_t].
-  */
-#define seq_nt16_char2idx(c) (seq_nt16_table[c])
-
-/*@abstract  Convert a char (A/C/G/T/N) to the index (0-4) of A/C/G/T/G.
-@param c     The base char [int8_t].
-@return      Index in the 'ACGTN' for the base [int8_t].
-  */
-#define seq_nt16_char2int(c) (seq_nt16_idx2int(seq_nt16_char2idx(c)))
-
-static char csp_nt5_str[] = "ACGTN";
-
-/*@abstract  Convert index in "ACGTN" to a letter.
-@param i     Index in "ACGTN"
-@return      A letter.
- */
-#define seq_nt16_int2char(i) (csp_nt5_str[i])
-
-/*@abstract  Convert 4-bit integer to a letter. seq_nt16_str is declared in htslib/hts.h 
-@param i     A 4-bit integer returned by bam_seqi(), standing for the index in the seq_nt16_str [int8_t].
-@return      A letter standing for base char [int8_t].
- */
-#define bam_seq_idx2base(i) (seq_nt16_str[i])
-
-/*@abstract  Convert raw qual score stored in bam1_t to qual char.
-@param i     Raw qual score stored in bam1_t [int8_t].
-@return      Qual char [int8_t].
-  */
-#define bam_seq_qual2char(i) ((i) + 33)
-
-/*@abstract   Get the content of bam aux tag of 'Z'/'H' (string) type.
-@param b      Pointer to the bam1_t structure.
-@param tag    Pointer to the bam aux tag.
-@return       NULL if donot contain the tag or corrupted data, pointer to the content otherwise.
-
-@note   1. To speed up, the caller should guarantee parameters b and tag are valid. 
-        2. The data of the pointer returned by this function is part of bam1_t, so do not double free!
- */
-static inline char* get_bam_aux_str(bam1_t *b, const char tag[2]) {
-    uint8_t *data;
-    if (NULL == (data = bam_aux_get(b, tag))) { return NULL; }
-    return bam_aux2Z(data);
-}
-
-/*
-* File API
- */
-
-/*@abstract  Packing the common bam file related pointers into a structure. */
-typedef struct {
-    htsFile *fp;
-    sam_hdr_t *hdr;   // hdr is needed by sam_read1().
-    hts_idx_t *idx;
-} csp_bam_fs;
-
-/*@abstract  Create a csp_bam_fs structure.
-@return  Pointer to the csp_bam_fs structure if success, NULL otherwise.
-
-@note    The pointer returned successfully by csp_bam_fs_init() should be freed
-         by csp_bam_fs_destroy() when no longer used.
- */
-static inline csp_bam_fs* csp_bam_fs_init(void) { return (csp_bam_fs*) calloc(1, sizeof(csp_bam_fs)); }
-
-static inline void csp_bam_fs_destroy(csp_bam_fs *p) {
-    if (p) {
-        if (p->idx) { hts_idx_destroy(p->idx); }
-        if (p->hdr) { sam_hdr_destroy(p->hdr); }
-        if (p->fp)  { hts_close(p->fp); }
-        free(p);
-    }
-}
-
-/*@abstract  Build the csp_bam_fs structure.  
-@param fn    Filename of the bam/sam/cram.
-@param ret   Pointer to the state.
-             0 if success, -1 for memory error, -2 for open/parse sam file error.
-@return      The pointer to the csp_bam_fs structure if success, NULL otherwise.
-
-@note        The pointer returned successfully by csp_bam_fs_build() should be freed
-             by csp_bam_fs_destroy() when no longer used.
-*/
-static inline csp_bam_fs* csp_bam_fs_build(const char *fn, int *ret) {
-    csp_bam_fs *p;
-    if (NULL == fn) { *ret = -1; return NULL; }
-    if (NULL == (p = csp_bam_fs_init())) { *ret = -1; return NULL; }
-    if (NULL == (p->fp = hts_open(fn, "rb"))) { *ret = -2; goto fail; }
-    if (NULL == (p->hdr = sam_hdr_read(p->fp))) { *ret = -2; goto fail; }
-    if (NULL == (p->idx = sam_index_load(p->fp, fn))) { *ret = -2; goto fail; }
-    *ret = 0;
-    return p;
-  fail:
-    csp_bam_fs_destroy(p);
-    return NULL;		
-}
-
-static inline int csp_bam_fs_reset(csp_bam_fs *p, const char *fn) {
-    if (NULL == p) { return -1; }
-    if (p->idx) { hts_idx_destroy(p->idx); }
-    if (p->hdr) { sam_hdr_destroy(p->hdr); }
-    if (p->fp)  { hts_close(p->fp); }		
-    if (NULL == (p->fp = hts_open(fn, "rb"))) { return -1; }
-    if (NULL == (p->hdr = sam_hdr_read(p->fp))) { return -1; }
-    if (NULL == (p->idx = sam_index_load(p->fp, fn))) { return -1; }
-    return 0;
-}
-
-/* let gzgets() to be compatible with kgetline(). */
-static inline char* csp_gzgets(char *buf, int len, gzFile fp) { return gzgets(fp, buf, len); }
-static inline int csp_bgzf_gets(BGZF *fp, kstring_t *s) {
-    return bgzf_getline(fp, '\n', s) <= -1 ? -1 : 0;
-}
-
-#if (CSP_FS_ZIP_METHOD == CSP_FS_BGZIP)
-    #define CSP_FS_ZFILE BGZF*
-    #define csp_fs_zopen(fn, fm) bgzf_open(fn, fm)
-    #define csp_fs_zclose(zfp) bgzf_close(zfp)
-    #define csp_fs_zread(zfp, buf, len) bgzf_read(zfp, buf, len)
-    #define csp_fs_zgetln(zfp, ks) csp_bgzf_gets(zfp, ks)
-    #define csp_fs_zwrite(zfp, buf, len) bgzf_write(zfp, buf, len)
-#else
-    #define CSP_FS_ZFILE gzFile
-    #define csp_fs_zopen(fn, fm) gzopen(fn, fm)
-    #define csp_fs_zclose(zfp) gzclose(zfp)
-    #define csp_fs_zread(zfp, buf, len) gzread(zfp, buf, len)
-    #define csp_fs_zgetln(zfp, ks) kgetline(ks, (kgets_func*) csp_gzgets, zfp)
-    #define csp_fs_zwrite(zfp, buf, len) gzwrite(zfp, buf, len)
-#endif
-
-/*@abstract    Structure used for outputed files with Simple Output Buffer.
-@param fn      Filename.
-@param fm      Filemode.
-@param zfp     File pointer for zipped mode.
-@param fp      File pointer for non-zipped mode.
-@param is_zip  If the outputed file should be zipped.
-@param is_tmp  If the outputed file is a tmp file. It's only used by mtx file and has no effect on I/O.
-@param is_open If the outputed file is open.
-@param buf     Mimic Output Buffer.
-@param bufsize Size of buffer.
-@note          1. The @p fn should be valid pointer coming from strdup().
-               2. The @p fm points to const string, so do not free it!
-               3. Output buffer is inside the structure.
-@TODO  Add is_error to save the state that if I/O has error.
- */
-typedef struct {
-    char *fn;
-    char *fm;
-    CSP_FS_ZFILE zfp;
-    FILE *fp;
-    uint8_t is_zip, is_tmp, is_open;
-    kstring_t ks, *buf;
-    size_t bufsize;
-} csp_fs_t;
-
-/*@abstract  Initialize the csp_fs_t structure.
-@param       Void.
-@return      Pointer of csp_fs_t if success, NULL otherwise.
- */
-static inline csp_fs_t* csp_fs_init(void) { 
-    csp_fs_t *p = (csp_fs_t*) calloc(1, sizeof(csp_fs_t));
-    if (p) { p->buf = &p->ks; p->bufsize = 1048576; }   // p->ks has been initialized after calling calloc(). Double initializing will cause error.
-    return p;
-}
-
-static inline void csp_fs_destroy(csp_fs_t* p) {
-    if (p) {
-        if (p->is_open) {
-            if (p->is_zip) { csp_fs_zclose(p->zfp); p->zfp = NULL; }
-            else { fclose(p->fp); p->fp = NULL; }
-        }
-        ks_free(p->buf);
-        free(p->fn); free(p);
-    }
-}
-
-static inline void csp_fs_set_bufsize(csp_fs_t *p, size_t bufsize) { p->bufsize = bufsize; }
-
-/* If the csp_fs_t is open. 0:no; 1:yes. */
-static inline int csp_fs_isopen(csp_fs_t *p) { return p->is_open; }
-
-/*@abstract  Open csp_fs_t for reading or writing.
-@param p     Pointer of csp_fs_t.
-@param mode  Open mode as in fopen(). if NULL, use default mode inside csp_fs_t.
-@return      1 if success, 0 if already open, -1 if error. 
- */
-static inline int csp_fs_open(csp_fs_t *p, char *mode) {
-    if (p->is_open) { return 0; }
-    char *fm = mode ? mode : p->fm;
-    if (p->is_zip) {
-        if (NULL == (p->zfp = csp_fs_zopen(p->fn, fm))) { return -1; }
-        else { p->is_open = 1; return 1; }
-    } else if (NULL == (p->fp = fopen(p->fn, fm))) {
-        return -1;
-    } else { p->is_open = 1; return 1; }
-}
-
-/*@abstract  Read from csp_fs_t without Input buffer.
-@param p     Pointer of csp_fs_t.
-@param buf   Buffer where the read content will be pushed into.
-@param len   Size of content to be read.
-@return      Value returned as gzread/fread. */
-static inline ssize_t csp_fs_read(csp_fs_t *p, char *buf, size_t len) {
-    return p->is_zip ? csp_fs_zread(p->zfp, buf, len) : fread(buf, 1, len, p->fp);
-}
-
-/*@abstract  Get a line from csp_fs_t without Input buffer.
-@param p     Pointer of csp_fs_t.
-@param s     Pointer of kstring_t which saves the line.
-@return      Value returned as kgetline(). 0 if success, EOF if end-of-file or error.
-
-@note        kgetline() is defined in htslib/kstring.h 
-*/
-static inline int csp_fs_getln(csp_fs_t *p, kstring_t *s) {
-    return p->is_zip ? csp_fs_zgetln(p->zfp, s) : kgetline(s, (kgets_func*) fgets, p->fp);
-}
-
-/*@abstract  Output functions below are like kxxx() functions in htslib/kstring.h with Output buffer.
-@return      Value returned are like in fxxx()/kxxx() functions (fputc()/kputc() etc.), which are 
-             the real size outputed if success, EOF if error. */
-
-/*@abstract  Flush buffer to stream.
-@param p     Pointer of csp_fs_t.
-@return      0 if success, EOF otherwise.
- */
-static inline int csp_fs_flush(csp_fs_t *p) {
-    ssize_t l, l0 = ks_len(p->buf);
-    l = p->is_zip ? csp_fs_zwrite(p->zfp, ks_str(p->buf), ks_len(p->buf)) : fwrite(ks_str(p->buf), 1, ks_len(p->buf), p->fp);
-    ks_clear(p->buf);
-    return l == l0 ? 0 : EOF;
-}
-
-static inline int csp_fs_printf(csp_fs_t *p, const char *fmt, ...) {
-    va_list ap;
-    int l;
-    va_start(ap, fmt);
-    l = kvsprintf(p->buf, fmt, ap);
-    va_end(ap);
-    if (ks_len(p->buf) >= p->bufsize && csp_fs_flush(p) < 0) { return EOF; }
-    return l;
-}
-
-static inline int csp_fs_putc(int c, csp_fs_t *p) { 
-    int l;
-    l = kputc(c, p->buf);
-    if (ks_len(p->buf) >= p->bufsize && csp_fs_flush(p) < 0) { return EOF; }
-    return l;
-}
-
-static inline int csp_fs_putc_(int c, csp_fs_t *p) { 
-    int l;
-    l = kputc_(c, p->buf);
-    if (ks_len(p->buf) >= p->bufsize && csp_fs_flush(p) < 0) { return EOF; }
-    return l;
-}
-
-static inline int csp_fs_puts(const char *s, csp_fs_t *p) { 
-    int l;
-    l = kputs(s, p->buf); 
-    if (ks_len(p->buf) >= p->bufsize && csp_fs_flush(p) < 0) { return EOF; }
-    return l;
-}
-
-static inline int csp_fs_write(csp_fs_t *p, char *buf, size_t len) { 
-    int l; 
-    l = kputsn(buf, len, p->buf);
-    if (ks_len(p->buf) >= p->bufsize && csp_fs_flush(p) < 0) { return EOF; }
-    return l;
-}
-
-/*@abstract  Close csp_fs_t but does not destroy it.
-@param p     Pointer of csp_fs_t.
-@return      0 if success, EOF otherwise.
-
-@note        Even fail, the csp_fs_t will still be set to not open.
- */
-static inline int csp_fs_close(csp_fs_t *p) {
-    int ret = 0;
-    if (p->is_open) {
-        if (ks_len(p->buf) && csp_fs_flush(p) < 0) { ret = EOF; } // only for write mode.
-        if (p->is_zip) { csp_fs_zclose(p->zfp); p->zfp = NULL; }
-        else { fclose(p->fp); p->fp = NULL; }
-        p->is_open = 0;
-    } 
-    return ret;
-}
-
-/*@abstract   Remove file.
-@param p      Pointer of csp_fs_t to be removed.
-@return       1 if success, 0 if file does not exist, -1 if failure.
- */
-static inline int csp_fs_remove(csp_fs_t *p) {
-    if (0 != access(p->fn, F_OK)) { return 0; }
-    if (remove(p->fn) < 0) { return -1; }
-    return 1;
-}
-
-/*@abstract   Remove all files in array.
-@param fs     Pointer of array of csp_fs_t to be removed.
-@param n      Size of array.
-@return       Num of files have been removed if no error, -1 otherwise.
- */
-static inline int csp_fs_remove_all(csp_fs_t **fs, const int n) {
-    int i, j, ret;
-    for (i = 0, j = 0; i < n; i++) {
-        if ((ret = csp_fs_remove(fs[i])) < 0) { return -1; }
-        else { j += ret; }
-    }
-    return j;
-}
-
-/* 
-* SNP List API
-*/
-/*@abstruct    The basic data structure to store SNP info.
-@param chr     Name of chr.
-@param pos     0-based coordinate in the reference sequence.
-@param ref     Ref base (a letter). 0 means no ref in the input SNP file for the pos.
-@param alt     Alt base (a letter). 0 means no alt in the input SNP file for the pos.
- */
-typedef struct {
-    char *chr;   
-    hts_pos_t pos; 
-    int8_t ref, alt;
-} csp_snp_t;
-
-/*@abstract  Initilize the csp_snp_t structure.
-@return      Pointer to the csp_snp_t structure if success, NULL if failure.
-@note        The pointer returned successfully by csp_snp_init() should be freed
-             by csp_snp_destroy() when no longer used.
- */
-static inline csp_snp_t* csp_snp_init(void) { return (csp_snp_t*) calloc(1, sizeof(csp_snp_t)); }
-
-static inline void csp_snp_destroy(csp_snp_t *p) { 
-    if (p) { free(p->chr); free(p); } 
-}
-
-static inline void csp_snp_reset(csp_snp_t *p) {
-    if (p) { free(p->chr); memset(p, 0, sizeof(csp_snp_t)); }
-}
-
-/*@abstract  A list containing of several pointers to the csp_snp_t structure.
-@param a     Array of csp_snp_t* pointers.
-@param n     The next pos of the unused element of the array.
-@param m     Size of the whole array.
-
-@note        The csp_snplist_t structure should be freed by csp_snplist_destroy() when no longer used.
-             csp_snplist_init() function should be called immediately after the structure was created.
-
-@example (A simple example from kvec.h)
-    kvec_t(int) array;
-    kv_init(array);
-    kv_push(int, array, 10); // append
-    kv_A(array, 20) = 4;
-    kv_destroy(array);
- */
-typedef kvec_t(csp_snp_t*) csp_snplist_t;   /* kvec_t from kvec.h */
-#define csp_snplist_init(v) kv_init(v)
-#define csp_snplist_resize(v, size) kv_resize(csp_snp_t*, v, size)
-#define csp_snplist_push(v, x) kv_push(csp_snp_t*, v, x)
-#define csp_snplist_A(v, i) kv_A(v, i)
-#define csp_snplist_size(v) kv_size(v)
-#define csp_snplist_max(v) kv_max(v)
-#define csp_snplist_destroy(v) {														\
-    size_t __j;																			\
-    for (__j = 0; __j < csp_snplist_size(v); __j++) csp_snp_destroy(csp_snplist_A(v, __j));	\
-    kv_destroy(v);																		\
-}
-
-/*@abstract    Extract SNP info from bcf/vcf file.
-@param fn      Filename of bcf/vcf.
-@param pl      Pointer to client data used to store the extracted SNP info.
-@param ret     Pointer to store running state. 0 if success, -1 otherwise.
-@param print_skip  If print the skipped SNPs. 1, yes, 0, no.
-@return        Num of elements successfully added into the snplist.
-
-@note          If length of Ref or Alt is larger than 1, then the SNP would be skipped.
-               If length of Ref or Alt is 0, then their values would be infered during pileup.
- */
-static size_t get_snplist_from_vcf(const char *fn, csp_snplist_t *pl, int *ret, int print_skip) {
-    htsFile *fp = NULL;
-    bcf_hdr_t *hdr = NULL;
-    bcf1_t *rec = NULL;
-    csp_snp_t *ip = NULL;
-    size_t l, m, n = 0;  /* use n instead of kv_size(*pl) in case that pl is not empty at the beginning. */
-    int r;
-    *ret = -1;
-    if (NULL == fn || NULL == pl) { return 0; }
-    if (NULL == (fp = hts_open(fn, "rb"))) { fprintf(stderr, "[E::%s] could not open '%s'\n", __func__, fn); return 0; }
-    if (NULL == (hdr = bcf_hdr_read(fp))) { fprintf(stderr, "[E::%s] could not read header for '%s'\n", __func__, fn); goto fail; }
-    if (NULL == (rec = bcf_init1())) { fprintf(stderr, "[E::%s] could not initialize the bcf structure.\n", __func__); goto fail; }
-    for (m = 1; (r = bcf_read1(fp, hdr, rec)) >= 0; m++) {
-        if (NULL == (ip = csp_snp_init())) { 
-            fprintf(stderr, "[E::%s] could not initialize the csp_snp_t structure.\n", __func__); 
-            goto fail; 
-        }
-        ip->chr = safe_strdup(bcf_hdr_id2name(hdr, rec->rid));
-        if (NULL == ip->chr) {
-            if (print_skip) { fprintf(stderr, "[W::%s] skip No.%ld SNP: could not get chr name.\n", __func__, m); }
-            csp_snp_destroy(ip);
-            continue;
-        } ip->pos = rec->pos;
-        bcf_unpack(rec, BCF_UN_STR);
-        if (rec->n_allele > 0) {
-            if (1 == (l = strlen(rec->d.allele[0]))) { ip->ref = rec->d.allele[0][0]; }
-            else if (l > 1) { 
-                if (print_skip) { fprintf(stderr, "[W::%s] skip No.%ld SNP: ref_len > 1.\n", __func__, m); }
-                csp_snp_destroy(ip); 
-                continue; 
-            } // else: do nothing. keep ref = 0, 0 is its init value.               
-            if (2 == rec->n_allele) {
-                if (1 == (l = strlen(rec->d.allele[1]))) { ip->alt = rec->d.allele[1][0]; }
-                else if (l > 1) {
-                    if (print_skip) { fprintf(stderr, "[W::%s] skip No.%ld SNP: alt_len > 1.\n", __func__, m); }
-                    csp_snp_destroy(ip);
-                    continue; 					
-                } // else: do nothing. keep alt = 0, 0 is its init value.
-            } else if (rec->n_allele > 2) {
-                if (print_skip) { fprintf(stderr, "[W::%s] skip No.%ld SNP: n_allele > 2.\n", __func__, m); }
-                csp_snp_destroy(ip);
-                continue;                 
-            } // else: keep alt = 0.
-        } // else: do nothing. keep ref = alt = 0, 0 is their init values.
-        csp_snplist_push(*pl, ip);
-        n++;
-    }
-    if (-1 == r) { // end of bcf file.
-        csp_snplist_resize(*pl, csp_snplist_size(*pl));
-    } else { 
-        fprintf(stderr, "[E::%s] error when parsing '%s'\n", __func__, fn); 
-        goto fail; 
-    }
-    bcf_destroy1(rec);
-    bcf_hdr_destroy(hdr);
-    hts_close(fp);
-    *ret = 0;
-    return n;
-  fail:
-    if (rec) bcf_destroy1(rec);
-    if (hdr) bcf_hdr_destroy(hdr);
-    if (fp) hts_close(fp);
-    return n;
-}
-#define get_snplist(fn, pl, ret, print_skip) get_snplist_from_vcf(fn, pl, ret, print_skip)
-
-/* 
-* Thread API
-*/
-/*@abstract    The data structure used as thread-func parameter.
-@param gs      Pointer to the global_settings structure.
-@param n       Pos of next element in the snp-list/chrom-list to be used by certain thread.
-@param m       Total size of elements to be used by certain thread, must not be changed.
-@param i       Id of the thread data.
-@param ret     Running state of the thread.
-@param ns      Num of SNPs that passed all filters.
-@param nr_*    Num of records for each output matrix file. 
-@param out_*   Pointers of output files.
- */
-typedef struct {
-    global_settings *gs;
-    size_t m, n;   // for snp-list or chrom-list.
-    int i;
-    int ret;
-    size_t ns, nr_ad, nr_dp, nr_oth;
-    csp_fs_t *out_mtx_ad, *out_mtx_dp, *out_mtx_oth, *out_vcf_base, *out_vcf_cells;
-} thread_data;
-
-/*@abstract  Create the thread_data structure.
-@return      Pointer to the structure if success, NULL otherwise.
-@note        The pointer returned successfully by thdata_init() should be freed
-             by thdata_destroy() when no longer used.
- */
-static inline thread_data* thdata_init(void) { return (thread_data*) calloc(1, sizeof(thread_data)); }
-
-static inline void thdata_destroy(thread_data *p) { free(p); }
-
-static inline void thdata_print(FILE *fp, thread_data *p) {
-    fprintf(fp, "\tm = %ld, n = %ld\n", p->m, p->n);
-    fprintf(fp, "\ti = %d, ret = %d\n", p->i, p->ret);
-}
+#include "jnumeric.h"
+#include "jfile.h"
+#include "jmemory.h"
+#include "config.h"
 
 /*
 * Pileup and MPileup API
@@ -745,14 +202,14 @@ typedef khash_t(ug) csp_map_ug_t;
 #define csp_map_ug_end(h) kh_end(h)
 #define csp_map_ug_size(h) kh_size(h)
 #define csp_map_ug_reset(h) kh_clear(ug, h)
-#define csp_map_ug_destroy(h) {														\
-    if (h) {																			\
-        csp_map_ug_iter __k;															\
-        for (__k = csp_map_ug_begin(h); __k != csp_map_ug_end(h); __k++) { 				\
+#define csp_map_ug_destroy(h) {								\
+    if (h) {										\
+        csp_map_ug_iter __k;								\
+        for (__k = csp_map_ug_begin(h); __k != csp_map_ug_end(h); __k++) { 			\
             if (csp_map_ug_exist(h, __k)) csp_list_uu_destroy(csp_map_ug_val(h, __k));		\
-        }																				\
-        kh_destroy(ug, h);																\
-    }																					\
+        }											\
+        kh_destroy(ug, h);									\
+    }												\
 }
 
 /* Struct csp_list_qu_t APIs 
@@ -965,18 +422,18 @@ static int csp_plp_str_vcf(csp_plp_t *p, kstring_t *s) {
     return 0;
 }
 
-static int csp_plp_to_vcf(csp_plp_t *p, csp_fs_t *s) {
-    if (p->tc <= 0) { csp_fs_puts(".:.:.:.:.:.", s); return 0; }
+static int csp_plp_to_vcf(csp_plp_t *p, jfile_t *s) {
+    if (p->tc <= 0) { jf_puts(".:.:.:.:.:.", s); return 0; }
     double gl[5];
     int i, m;
     double tmp = -10 / log(10);
     char *gt[] = {"0/0", "1/0", "1/1"};
     m = get_idx_of_max(cu_d, p->gl, 3);
-    csp_fs_puts(gt[m], s);
-    csp_fs_printf(s, ":%ld:%ld:%ld:", p->ad, p->dp, p->oth);
+    jf_puts(gt[m], s);
+    jf_printf(s, ":%ld:%ld:%ld:", p->ad, p->dp, p->oth);
     for (i = 0; i < p->ngl; i++) { gl[i] = p->gl[i] * tmp; }
     if (join_arr_to_str(cu_d, gl, p->ngl, ',', "%.0f", s->buf) < p->ngl) { return -1; }  // TODO: use internal buf directly is not good.
-    csp_fs_putc_(':', s);
+    jf_putc_(':', s);
     if (join_arr_to_str(cu_s, p->bc, 5, ',', "%ld", s->buf) < 5) { return -1; }
     return 0;
 }
@@ -999,22 +456,22 @@ typedef khash_t(sg) csp_map_sg_t;
 #define csp_map_sg_begin(h) kh_begin(h)
 #define csp_map_sg_end(h) kh_end(h)
 #define csp_map_sg_size(h) kh_size(h)
-#define csp_map_sg_destroy(h) {														\
-    if (h) {																		\
-        csp_map_sg_iter __k;														\
-        for (__k = csp_map_sg_begin(h); __k != csp_map_sg_end(h); __k++) { 				\
+#define csp_map_sg_destroy(h) {								\
+    if (h) {											\
+        csp_map_sg_iter __k;									\
+        for (__k = csp_map_sg_begin(h); __k != csp_map_sg_end(h); __k++) { 			\
             if (csp_map_sg_exist(h, __k)) csp_plp_destroy(csp_map_sg_val(h, __k)); 	\
-        }																			\
-        kh_destroy(sg, h);															\
-    }																				\
+        }										\
+        kh_destroy(sg, h);									\
+    }												\
 }
-#define csp_map_sg_reset_val(h) {														\
-    if (h) {																			\
-        csp_map_sg_iter __k;															\
-        for (__k = csp_map_sg_begin(h); __k != csp_map_sg_end(h); __k++) {				\
+#define csp_map_sg_reset_val(h) {								\
+    if (h) {											\
+        csp_map_sg_iter __k;									\
+        for (__k = csp_map_sg_begin(h); __k != csp_map_sg_end(h); __k++) {			\
             if (csp_map_sg_exist(h, __k)) csp_plp_reset(csp_map_sg_val(h, __k)); 		\
-        }																				\
-    }																					\
+        }											\
+    }												\
 }
 
 /*@abstract  The structure stores the stat info of all sample groups for certain query pos.
@@ -1174,10 +631,10 @@ static inline int csp_mplp_str_vcf(csp_mplp_t *mplp, kstring_t *s) {
     return 0;
 }
 
-static inline int csp_mplp_to_vcf(csp_mplp_t *mplp, csp_fs_t *s) {
+static inline int csp_mplp_to_vcf(csp_mplp_t *mplp, jfile_t *s) {
     int i;
     for (i = 0; i < mplp->nsg; i++) {
-        csp_fs_putc_('\t', s);
+        jf_putc_('\t', s);
         if (csp_plp_to_vcf(csp_map_sg_val(mplp->hsg, mplp->hsg_iter[i]), s) < 0) { return -1; }
     } //s->s[--(s->l)] = '\0';    /* s->l could not be negative unless no csp_plp_t(s) are printed to s->s. */
     return 0;
@@ -1225,18 +682,18 @@ static inline int csp_mplp_str_mtx_tmp(csp_mplp_t *mplp, kstring_t *ks_ad, kstri
     return 0; 
 }
 
-static int csp_mplp_to_mtx(csp_mplp_t *mplp, csp_fs_t *fs_ad, csp_fs_t *fs_dp, csp_fs_t *fs_oth, size_t idx) {
+static int csp_mplp_to_mtx(csp_mplp_t *mplp, jfile_t *fs_ad, jfile_t *fs_dp, jfile_t *fs_oth, size_t idx) {
     csp_plp_t *plp;
     int i;
     for (i = 1; i <= mplp->nsg; i++) {
         plp = csp_map_sg_val(mplp->hsg, mplp->hsg_iter[i - 1]);
-        if (plp->ad) fs_ad->is_tmp ? csp_fs_printf(fs_ad, "%d\t%ld\n", i, plp->ad) : csp_fs_printf(fs_ad, "%ld\t%d\t%ld\n", idx, i, plp->ad);
-        if (plp->dp) fs_dp->is_tmp ? csp_fs_printf(fs_dp, "%d\t%ld\n", i, plp->dp) : csp_fs_printf(fs_dp, "%ld\t%d\t%ld\n", idx, i, plp->dp);
-        if (plp->oth) fs_oth->is_tmp ? csp_fs_printf(fs_oth, "%d\t%ld\n", i, plp->oth) : csp_fs_printf(fs_oth, "%ld\t%d\t%ld\n", idx, i, plp->oth);      
+        if (plp->ad) fs_ad->is_tmp ? jf_printf(fs_ad, "%d\t%ld\n", i, plp->ad) : jf_printf(fs_ad, "%ld\t%d\t%ld\n", idx, i, plp->ad);
+        if (plp->dp) fs_dp->is_tmp ? jf_printf(fs_dp, "%d\t%ld\n", i, plp->dp) : jf_printf(fs_dp, "%ld\t%d\t%ld\n", idx, i, plp->dp);
+        if (plp->oth) fs_oth->is_tmp ? jf_printf(fs_oth, "%d\t%ld\n", i, plp->oth) : jf_printf(fs_oth, "%ld\t%d\t%ld\n", idx, i, plp->oth);      
     }
-    if (fs_ad->is_tmp) csp_fs_putc('\n', fs_ad);
-    if (fs_dp->is_tmp) csp_fs_putc('\n', fs_dp);
-    if (fs_oth->is_tmp) csp_fs_putc('\n', fs_oth);
+    if (fs_ad->is_tmp) jf_putc('\n', fs_ad);
+    if (fs_dp->is_tmp) jf_putc('\n', fs_dp);
+    if (fs_oth->is_tmp) jf_putc('\n', fs_oth);
     return 0; 
 }
